Relocate and rename main component file
 To fit what's in the "master" branch
