+++ src/index.js
@@ -1,2 +1,2 @@
-import VueTimepicker from './components/VueTimepicker'
+import VueTimepicker from './vue-timepicker'
 export default VueTimepicker
